# multas-backend/docs/API_OPENAPI.yaml
openapi: 3.1.0
info:
  title: PMT Multas API - TANDAS 0–8, 10 y 11
  version: 1.2.0
  description: Contrato de autenticación, operación municipal, finanzas, solvencias, analítica, notificaciones y migración histórica controlada.
servers:
  - url: /api/v1
...
paths:
...
  /agents:
    get:
      tags: [Administration]
      security: [{ sessionCookie: [] }]
      summary: Lista agentes PMT
      responses: { '200': { description: Página de agentes } }
    post:
      tags: [Administration]
      security: [{ sessionCookie: [] }]
      summary: Registra un agente único relacionado con un usuario activo
      responses: { '201': { description: Agente creado }, '409': { $ref: '#/components/responses/Error' } }
  /agents/{id}:
    get:
      tags: [Administration]
      security: [{ sessionCookie: [] }]
      summary: Consulta un agente PMT
      responses: { '200': { description: Agente PMT }, '404': { $ref: '#/components/responses/Error' } }
    patch:
      tags: [Administration]
      security: [{ sessionCookie: [] }]
      summary: Actualiza la relación operativa de un agente
      responses: { '204': { description: Agente actualizado }, '409': { $ref: '#/components/responses/Error' } }
...
  /dashboard:
    get:
      tags: [Dashboard]
      security: [{ sessionCookie: [] }]
      summary: Devuelve KPI y series agregadas por MySQL
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: departmentId, in: query, schema: { type: integer } }
      responses: { '200': { description: Indicadores, gráficas y dependencias }, '403': { $ref: '#/components/responses/Error' }, '422': { $ref: '#/components/responses/Error' } }
  /reports/{type}:
    get:
      tags: [Reports]
      security: [{ sessionCookie: [] }]
      summary: Consulta un reporte paginado con filtros
      parameters:
        - { name: type, in: path, required: true, schema: { type: string, enum: [infractions, collection, cash, payments-reversals, reconciliation, aging, adjustments-exemptions, appeals, solvencies, agent-activity, audit] } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: departmentId, in: query, schema: { type: integer } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses: { '200': { description: Filas y metadatos de paginación }, '403': { $ref: '#/components/responses/Error' } }
  /reports/{type}/export.csv:
    get:
      tags: [Reports]
      security: [{ sessionCookie: [] }]
      summary: Exporta hasta 10000 filas por lotes en CSV UTF-8
      parameters:
        - { name: type, in: path, required: true, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses: { '200': { description: Archivo CSV }, '403': { $ref: '#/components/responses/Error' } }
  /reports/dashboard-summary.pdf:
    get:
      tags: [Reports]
      security: [{ sessionCookie: [] }]
      summary: Exporta el resumen administrativo en PDF
      responses: { '200': { description: Archivo PDF }, '403': { $ref: '#/components/responses/Error' } }
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pmt_session
  responses:
    Error:
      description: Error estable con requestId
      content:
        application/json:
          schema:
            type: object
            required: [error, meta]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: { type: string }
                  message: { type: string }
              meta:
                type: object
                required: [requestId]
                properties:
                  requestId: { type: string, format: uuid }