# multas-backend/src/modules/agents/http/agent.routes.ts
import { Router } from "express";
import type { ResultSetHeader, RowDataPacket } from "mysql2/promise";
import { z } from "zod";
import type { AppContainer } from "../../../bootstrap/container.js";
import { authenticate } from "../../../shared/http/authenticate.js";
import { authorize } from "../../../shared/http/authorize.js";
import { HttpError } from "../../../shared/http/HttpError.js";
import { getRequestId } from "../../../shared/http/request-context.js";
import { isDuplicateKey, recordOperation } from "../../../shared/http/operations.js";

const inputSchema=z.object({userId:z.coerce.number().int().positive(),badgeNumber:z.string().trim().min(2).max(50),hiredAt:z.iso.date().nullable().optional()});
const updateSchema=inputSchema.partial().refine(value=>Object.keys(value).length>0);
const listSchema=z.object({search:z.string().trim().max(100).default(""),status:z.enum(["ACTIVE","SUSPENDED","INACTIVE"]).optional(),page:z.coerce.number().int().min(1).default(1),pageSize:z.coerce.number().int().min(1).max(100).default(20)});

export function createAgentRouter(container:AppContainer):Router{
 const router=Router();router.use(authenticate(container.authRepository,container.env));
 router.get("/",authorize("users.read",container.auditRepository),async(request,response,next)=>{try{const q=listSchema.parse(request.query);const clauses:string[]=[];const values:unknown[]=[];if(q.status){clauses.push("a.status=?");values.push(q.status);}if(q.search){clauses.push("(a.badge_number LIKE ? OR u.username LIKE ? OR u.first_name LIKE ? OR u.last_name LIKE ?)");const like=`%${q.search}%`;values.push(like,like,like,like);}const where=clauses.length?`WHERE ${clauses.join(" AND ")}`:"";const counts=await container.database.query<(RowDataPacket&{total:number})[]>(`SELECT COUNT(*) total FROM agents a JOIN users u ON u.id=a.user_id ${where}`,values);const rows=await container.database.query<RowDataPacket[]>(`SELECT a.id,a.user_id,a.badge_number,a.status,a.hired_at,a.created_at,a.updated_at,u.username,u.first_name,u.last_name,u.status user_status FROM agents a JOIN users u ON u.id=a.user_id ${where} ORDER BY a.badge_number LIMIT ? OFFSET ?`,[...values,q.pageSize,(q.page-1)*q.pageSize]);response.json({data:rows,meta:{page:q.page,pageSize:q.pageSize,total:counts[0]?.total??0,requestId:getRequestId()}});}catch(error){next(error);}});
 router.get("/:id",authorize("users.read",container.auditRepository),async(request,response,next)=>{try{const id=z.coerce.number().int().positive().parse(request.params["id"]);const rows=await container.database.query<RowDataPacket[]>("SELECT a.*,u.username,u.first_name,u.last_name,u.status user_status FROM agents a JOIN users u ON u.id=a.user_id WHERE a.id=?",[id]);if(!rows[0])throw new HttpError({code:"AGENT_NOT_FOUND",message:"Agente no encontrado.",statusCode:404});response.json({data:rows[0],meta:{requestId:getRequestId()}});}catch(error){next(error);}});
 router.post("/",authorize("users.create",container.auditRepository),async(request,response,next)=>{try{const input=inputSchema.parse(request.body);const users=await container.database.query<(RowDataPacket&{status:string})[]>("SELECT status FROM users WHERE id=?",[input.userId]);if(!users[0])throw new HttpError({code:"USER_NOT_FOUND",message:"Usuario no encontrado.",statusCode:404});if(users[0].status!=="ACTIVE")throw new HttpError({code:"AGENT_USER_INACTIVE",message:"El usuario del agente debe estar activo.",statusCode:409});const result=await container.database.query<ResultSetHeader>("INSERT INTO agents (user_id,badge_number,hired_at,status) VALUES (?,?,?,'ACTIVE')",[input.userId,input.badgeNumber,input.hiredAt??null]);await recordOperation(container,request,{action:"AGENT_CREATED",module:"agents",entityType:"agent",entityId:String(result.insertId)});response.status(201).json({data:{id:String(result.insertId)},meta:{requestId:getRequestId()}});}catch(error){if(isDuplicateKey(error))return next(new HttpError({code:"AGENT_DUPLICATE",message:"El usuario o número institucional ya está asociado a un agente.",statusCode:409}));next(error);}});
 router.patch("/:id",authorize("users.update",container.auditRepository),async(request,response,next)=>{try{const id=z.coerce.number().int().positive().parse(request.params["id"]);const input=updateSchema.parse(request.body);const cols:string[]=[];const vals:unknown[]=[];for(const [key,col] of [["userId","user_id"],["badgeNumber","badge_number"],["hiredAt","hired_at"]] as const)if(input[key]!==undefined){cols.push(`${col}=?`);vals.push(input[key]);}const result=await container.database.query<ResultSetHeader>(`UPDATE agents SET ${cols.join(",")} WHERE id=?`,[...vals,id]);if(!result.affectedRows)throw new HttpError({code:"AGENT_NOT_FOUND",message:"Agente no encontrado.",statusCode:404});await recordOperation(container,request,{action:"AGENT_UPDATED",module:"agents",entityType:"agent",entityId:String(id)});response.sendStatus(204);}catch(error){if(isDuplicateKey(error))return next(new HttpError({code:"AGENT_DUPLICATE",message:"El usuario o número institucional ya está asociado a un agente.",statusCode:409}));next(error);}});
 for(const [path,status,action] of [["activate","ACTIVE","AGENT_ACTIVATED"],["deactivate","INACTIVE","AGENT_DEACTIVATED"]] as const)router.post(`/:id/${path}`,authorize("users.change_status",container.auditRepository),async(request,response,next)=>{try{const id=z.coerce.number().int().positive().parse(request.params["id"]);const result=await container.database.query<ResultSetHeader>("UPDATE agents SET status=? WHERE id=?",[status,id]);if(!result.affectedRows)throw new HttpError({code:"AGENT_NOT_FOUND",message:"Agente no encontrado.",statusCode:404});await recordOperation(container,request,{action,module:"agents",entityType:"agent",entityId:String(id)});response.sendStatus(204);}catch(error){next(error);}});
 return router;
}